Split test data into separate files.
