+++ docker/docker-compose.yaml
@@ -4,7 +4,7 @@ version: "3.8"
 services:
   app:
     container_name: cksource-monitoring
-    build: 
+    build:
       context: ../
       dockerfile: ./docker/tests-runner/Dockerfile.dev
     env_file:
@@ -21,7 +21,7 @@ services:
     env_file:
       - ../.env
     environment:
-      GF_AUTH_ANONYMOUS_ENABLED: true 
+      GF_AUTH_ANONYMOUS_ENABLED: 'true'
     ports:
       - "3000:3000"
