+++ src/data/CERTIFICATES_DATA.ts
@@ -1,0 +1,78 @@
+/*
+ Copyright (c), CKSource Holding sp. z o.o. All rights reserved.
+ */
+
+import { OrganizationName } from './types.js';
+
+export const CERTIFICATES_DATA: Record<OrganizationName, string[]> = {
+	Tiugo: [
+		'https://tiugotech.com'
+	],
+	CKSource: [
+		'https://ckbox.io',
+		'https://cdn.ckbox.io',
+		'https://download.ckbox.io',
+		'https://cdn-source.ckbox.io',
+
+		'https://ckeditor.com',
+		'https://builder-api.ckeditor.com',
+		'https://cla.ckeditor.com',
+		'https://cdn.ckeditor.com',
+		'https://cdn-source.ckeditor.com',
+		'https://cke4.ckeditor.com',
+		'https://ckeditor-demo-preview-1.ckeditor.com',
+		'https://prev1.ckeditor.com',
+		'https://preview-blog.ckeditor.com',
+		'https://preview.ckeditor.com',
+		'https://svn.ckeditor.com',
+		'https://dev.ckeditor.com',
+		'https://d1.ckeditor.com',
+		'https://dashboard.ckeditor.com',
+		'https://sales.ckeditor.com',
+		'https://staging-orders.ckeditor.com',
+		'https://staging.ckeditor.com',
+		'https://docs-old.ckeditor.com',
+		'https://ckeditor.com/docs/',
+		'https://ckeditor.com/cke4/addons/plugins/all',
+
+		'https://ckeditor.dev/',
+		'https://develop.ckeditor.dev/',
+
+		'https://cke-cs.com',
+		'https://cke-cs-staging.com',
+		'https://cke-cs-dev.com',
+
+		'https://builder.ckeditor-dev.com',
+		'https://builder-demo.ckeditor-dev.com',
+		'https://builder-nightly.ckeditor-dev.com',
+
+		'https://cksource.com',
+		'https://a.cksource.com',
+		'https://c.cksource.com',
+		'https://cib.cksource.com',
+		'https://download.cksource.com',
+		'https://download-source.cksource.com',
+		'https://hanghub.cksource.com',
+		'https://host04.cksource.com',
+		'https://maven.cksource.com',
+		'https://outreach.cksource.com',
+		'https://preview.cksource.com',
+		'https://translations.cksource.com',
+
+		'https://fckeditor.net',
+		'https://mediawiki.fckeditor.net',
+
+		'https://onlinehtmleditor.dev',
+
+		'https://onlinemarkdowneditor.dev'
+	],
+	Tiny: [
+		'https://tiny.cloud',
+		'https://cla.tiny.cloud'
+	],
+	ButterCMS: [
+		'https://buttercms.com',
+		'https://api.buttercms.com',
+		'https://buttercms-ai-proxy.internal.cke-cs.com'
+	]
+};
