+++ src/data/DOMAINS_DATA.ts
@@ -1,0 +1,47 @@
+/*
+ Copyright (c), CKSource Holding sp. z o.o. All rights reserved.
+ */
+
+import { OrganizationName } from './types.js';
+
+export const DOMAINS_DATA: Record<OrganizationName, string[]> = {
+	Tiugo: [
+		'tiugotech.com'
+	],
+	CKSource: [
+		'ckbox.io',
+		'ckbox.cloud',
+		'ckdocs.io',
+		'ckeditor-cs.com',
+		'ckeditor-dev.com',
+		'ckeditor-staging.com',
+		'cke-cs.com',
+		'cke-cs-staging.com',
+		'cke-cs-dev.com',
+		'ckeditor.com',
+		'ckeditor.dev',
+		'ckeditor.net',
+		'ckeditor.org',
+		'ckfinder.com',
+		'ckfinder.net',
+		'cksource.com',
+		'collaborativeeditor.com',
+		'fckeditor.com',
+		'fckeditor.net',
+		'htmleditor.dev',
+		'markdowneditor.dev',
+		'onlinehtmleditor.dev',
+		'onlinemarkdowneditor.dev',
+		'realtimehtmleditor.net',
+		'richtexteditor.dev',
+		'wordtohtml.dev',
+		'wysiwygeditor.dev',
+		'wysiwygeditor.net'
+	],
+	Tiny: [
+		'tiny.cloud'
+	],
+	ButterCMS: [
+		'buttercms.com'
+	]
+};
