+++ src/data/types.ts
@@ -1,0 +1,5 @@
+/*
+ Copyright (c), CKSource Holding sp. z o.o. All rights reserved.
+ */
+
+export type OrganizationName = 'Tiugo' | 'CKSource' | 'Tiny' | 'ButterCMS';
